requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
opencv-python-headless==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
Pillow==10.0.1
redis==5.0.1
celery==5.3.4
websockets==11.0.2
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
boto3==1.34.84
django-storages==1.14.2
